The greatest commit in community app(change condition)

See merge request !31

+++ server/src/service/room/room.service.ts
@@ -115,7 +115,7 @@ export class RoomService {
     let room = this.rooms.find(r => r.id === index);
     const operation$ = Promise.resolve(true);
 
-    if (token) {
+    if (token && room) {
       this.playersBindService.removePlayers(room.token, token);
     }
